requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
asgiref==3.5.0
    # via django
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.12
    # via requests
coreapi==2.3.3
    # via
    #   django-rest-swagger
    #   openapi-codec
coreschema==0.0.4
    # via coreapi
django==4.0.3
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   djangorestframework
django-cors-headers==3.11.0
    # via -r requirements.in
django-rest-swagger==2.2.0
    # via -r requirements.in
djangorestframework==3.13.1
    # via
    #   -r requirements.in
    #   django-rest-swagger
idna==3.3
    # via requests
itypes==1.2.0
    # via coreapi
jinja2==3.1.1
    # via coreschema
markupsafe==2.1.1
    # via jinja2
openapi-codec==1.3.2
    # via django-rest-swagger
pytz==2022.1
    # via djangorestframework
pyyaml==6.0
    # via -r requirements.in
requests==2.27.1
    # via coreapi
simplejson==3.17.6
    # via django-rest-swagger
sqlparse==0.4.2
    # via django
uritemplate==4.1.1
    # via coreapi
urllib3==1.26.9
    # via requests
